# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This api is the backend api responsible for the whole Coo-e coordinator app.",
    "title": "Coordinator",
    "version": "",
    "x-stoplight": {
      "id": "haridasu/specs/coordinator"
    }
  },
  "host": "resources.coo-e.com:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/telluswho/senseOfCommunity": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Success",
                "profile": {
                  "createdAt": "2017-02-09T17:09:42.586Z",
                  "id": "86d59046-8a63-4473-a91b-af932b16f367",
                  "progress": 0,
                  "senseOfCommunity": {
                    "advice": 1,
                    "community": 1,
                    "connected": 1,
                    "createdAt": "2017-02-09T17:33:47.706Z",
                    "facultyCare": 1,
                    "friendSupport": 1,
                    "id": "04e80a2e-80d9-4479-9954-f7a2fe65349c",
                    "isolated": 1,
                    "likeFamily": 1,
                    "loan": 1,
                    "profileId": "86d59046-8a63-4473-a91b-af932b16f367",
                    "satisfied": 1,
                    "studentsCare": 1
                  },
                  "senseOfCommunityId": "04e80a2e-80d9-4479-9954-f7a2fe65349c",
                  "userId": "7cc5359a-8293-4117-a995-ff3c00830b5c"
                },
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Missing image data or image type.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Answers to survey in bulk"
      }
    },
    "/api/telluswho/sociability": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Success",
                "sociability": {
                  "createdAt": "2017-02-09T18:50:13.664Z",
                  "id": "5186d2d5-e7f1-4256-aa6e-8b0373145965",
                  "likePeople": 1,
                  "makingContacts": 1,
                  "mixSocially": 1,
                  "nervousAuthority": 1,
                  "oppositeSex": 1,
                  "peopleStimulating": 1,
                  "preferOthers": 1,
                  "sociallyAwkward": 1,
                  "talkToStrangers": 1,
                  "tenseWithPeople": 1,
                  "tenseWithStrangers": 1,
                  "uncomfortableParties": 1
                },
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Missing image data or image type.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Answers to survey in bulk"
      }
    },
    "/api/telluswho/wellbeing": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "message": "Success",
                "status": 200,
                "wellbeing": {
                  "createdAt": "2017-02-09T19:15:18.888Z",
                  "doNotHaveProud": 1,
                  "feelFailure": 1,
                  "feelIsolated": 1,
                  "feelLeftOut": 1,
                  "happyAtNjit": 1,
                  "highSelfEsteem": 1,
                  "id": "c6ee8ddd-657b-4d1e-9b3d-b1818148cf84",
                  "lackCompanionShip": 1,
                  "lifeExcellent": 1,
                  "lifeIdeal": 1,
                  "lifeSatisfied": 1,
                  "physicalHealth": 1,
                  "stayAtNjit": 1
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Missing image data or image type.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Answers to survey in bulk"
      }
    },
...
  },
...
}